third_party/blink/web_tests/platform/linux-chrome/external/wpt/clipboard-apis/async-unsanitized-plaintext-formats-write-read.tentative.https-expected.txt

This is a testharness.js-based test.
[FAIL] Verify write and read unsanitized content to the clipboard given standard and custom formats as input
  assert_equals: expected "Test text." but got ""
Harness: the test ran to completion.
